Fix parsing of numerics in expressions
Fixed issue in tokenization of expressions where numeric values with a decimal or in scientific notation were not being correctly tokenized, leading to potential errors in computed output. This should also fix the error in expanded display of expressions, as poor tokenization was the root cause.

Modified error log location to be local for built package.

Bumped version number; updated splash screen.